Pundit feedback
shycohen: Every scalable application needs queues. Generally speaking, if you don't need scalability, you don't need CQRS.
No change.

shycohen: It would be interesting to explain why CQRA made it harder to find issues. My experience shows otherwise: debugging issues is exactly the same as with a non separated system.
No change.

shycohen: Very true, but a combination of immediate feedback and eventual consistency makes throttling a non-issue in terms of propagation.Storage is a different mater, but it just eventually slows down the reads that trigger it, which goes back to eventual consistency.
No change.

shycohen: The problem is that you're reading from the event store. This is a workable, but not ideal implementation. My implemetation uses nothing but the built-in Azure (or AWS) mechanisms, so I don't need to build any infrastructure :)
No change.

shycohen: Ah! Thsi is why. This has little to do with CQRS. I recommend you change the above since it implies that the complexity is with the pattern.
No change.

shycohen: The link goes to a non-formatted page.
No change.

shycohen: Absolutely! Since the first thing that a user does after making a change is to view the change, latency is unacceptable. The commenting mechahism on this page is horrible because it doesn't give immediate feedback, and made me think at first that the application is not working (that it did not take my command).
No change.

shycohen: And throtling.
Done