Pundit feedback
mknz: This section and the next both start "As previously stated."
Modified second occurrence.

mknz: There is a lot of hedging in this sentence - "it is clear" ... "sometimes" ... "we are not saying"
Tightened this up

mknz: The use of tables is wrong here with regard to partitions. The transaction happens to multiple entities in a partition in a single table. The example at the end of Chapter 4 is correct.
Fixed this

mknz: Is there a reason the Storage Client retry policy mechanism wasn't used?http://blogs.msdn.com/b/windowsazurestorage/archive/2011/02/03/overview-of-retry-policies-in-the-windows-azure-storage-client-library.aspx
Added a callout.

mknz: It is probably worth pointing out explicitly that lines in red are event notifications. I was at first puzzled by the two lines marked 6 until I realized that.
Adjusted the text to talk a bit more about events rather than providing a key. We have a convention throughout the docs as to the style of lines/arrows that represent events and commands. Don't want to have to write this everywhere.

mknz: This section may be better merged with the Clarifying the terminology section since together they come across as somewhat repetitive. It is one of a number of places in this appendix where a hesitancy comes across as a lack of confidence in the documentation which is belied by the strength of the presentation.
Restructured this part.

mknz: This sentence needs clarification. On first reading it I thought it could also apply to queues but decided that the "multiple" allowed for multiple consumers in different subscriptions getting the message. All a bit too subtle..
No change - title says topic, sentence says topic, digram says topic...Quese were descibed in the previous section.