Pundit feedback
jdom: This is very old code in the RI. Order does not implement the IAggregateRoot interface, but the IEventSourced (which implies an event sourced aggregate root).The difference is that we no longer use an ORM for it, as expressed in the text.Perhaps we want to use the payment AR?
The code in the reference chapters is not intended to be from the RI - everything in the reference chapter can be read independently of the RI. No change.

jdom: what does it mean? business validity? if that's the case, no, we must accept every event as valid.If this means that the infrastructure validates that is not a duplicate or something, then it is fine, but it's very confusing.
REmoved - shouldn't have been there

jdom: Hmm, this is not completely true. Process managers send commands to other BCs (as they don't belong to any BC in particular).Also, there is the discussion about if the UI belongs to a BC in particular or it's outside of it, with different POVs
Added additional qualifying sentence

jdom: Version refers to the aggregate's version, and it's incremental. Every new event for a particular aggregate instance increments the version by 1.
Added sentence.

jdom: Version refers to the aggregate's version, and it's incremental. Every new event for a particular aggregate instance increments the version by 1.
Removed this sentence.

jdom: To avoid poison" messages to stall the processing of other messages."
Don't think this is necessary - no change

jdom: I'd add a dev remark that for simplicity in C#, they are implemented as DTOs, but they should be treated as immutable.
Changed

jdom: I'd add: the behavior is typically in one place, in the write model.
Done

jdom: I'd add: It favors anemic domain models, where the business logic or behavior is separated from the data, and potentially leaking in different layers.
NOt sure about this - does not necessarily have a domain model...

jdom:This is very old code which contained a leaky abstraction called out by Mark Seeman. The code changed since.
Changed

jdom: The current implementation is different from this, much more optimized and heavily asynchronous. Id mention this, but for simplicity in the docs, I'd keep this old version, which is much more readable.
Already states that it's not optimized.